+++ R/context.R
@@ -175,6 +175,7 @@ reply <- function(context,
 #'
 #' @inheritParams reply
 #' @inheritParams recv
+#' @param con a 'req' Socket, or a Context.
 #' @param data an object (if `send_mode = "raw"`, a vector).
 #' @param timeout \[default NULL\] integer value in milliseconds or NULL, which
 #'   applies a socket-specific default, usually the same as no timeout.
@@ -224,7 +225,7 @@ reply <- function(context,
 #'
 #' @export
 #'
-request <- function(context,
+request <- function(con,
                     data,
                     send_mode = c("serial", "raw"),
                     recv_mode = c("serial", "character", "complex", "double",
@@ -232,4 +233,4 @@ request <- function(context,
                     timeout = NULL,
                     cv = NULL,
                     msgid = NULL)
-  data <- .Call(rnng_request, context, data, send_mode, recv_mode, timeout, cv, msgid, environment())
+  data <- .Call(rnng_request, con, data, send_mode, recv_mode, timeout, cv, msgid, environment())